improvements

+++ iowait.sh
@@ -1,16 +1,23 @@
 #!/bin/bash
 ##Author Thanassis Zakopoulos
-##Usage  ./iowait.sh host_list "seconds to iterate" "iowait threshold"  &
+##Usage  ./iowait.sh host_list "seconds to iterate" "iowait threshold"  
 ##host_list file has to be of format "ip fqdn_hostname"
 
+## if file list is not given echo usage and exit
+if [ -z "$1" ]
+	then
+	echo 'Usage  ./iowait.sh host_list "seconds to iterate" "iowait threshold"'
+	exit 1
+fi
 ##if seconds period is not defined default will be 30
 SECONDS=$2
 if [ -z "$2" ]
 	then SECONDS=30
 fi
+##if iowait threshold is not defined default will be 1
 IOWAIT_THRESHOLD_PERCENTANCE=$3
 if [ -z "$3" ]
-then IOWAIT_THRESHOLD_PERCENTANCE=1
+	then IOWAIT_THRESHOLD_PERCENTANCE=1
 fi
 SCRIPT_HOME=/root/scripts/iowait
 YEAR=`date '+%Y'`
@@ -20,15 +27,18 @@ HOUR=`date '+%H'`
 MINUTE=`date '+%M'`
 SECOND=`date '+%S'`
 DATE=$YEAR.$MONTH.$DAY.$HOUR.$MINUTE.$SECOND
-TMP_LOG1=$SCRIPT_HOME/aaa.$DATE
-TMP_LOG2=$SCRIPT_HOME/bbb.$DATE
-TMP_LOG3=$SCRIPT_HOME/ccc.$DATE
+TMP_LOG1=$SCRIPT_HOME/tmp/aaa.$DATE
+TMP_LOG2=$SCRIPT_HOME/tmp/bbb.$DATE
+TMP_LOG3=$SCRIPT_HOME/tmp/ccc.$DATE
 LOG=$SCRIPT_HOME/log."$(echo $1)"
 
 cat /dev/null > $TMP_LOG1
 cat /dev/null > $TMP_LOG2
 cat /dev/null > $TMP_LOG3
-mv $LOG $LOG."$DATE"
+##before new execution if the script has been executed for a file list again backup old values of big iowaits
+if [ -f $LOG ]; then
+	mv $LOG $LOG."$DATE"
+fi
 cat /dev/null > $LOG
 
 function iowait () {
